everything working

+++ src/HealthComponent.cpp
@@ -1,6 +1,6 @@
 #include "HealthComponent.h"
 
-void HealthComponent_Add(HealthComponent* healthComponent, uint32 eid, uint8 startHealth) {
+void HealthComponent_Add(HealthComponent* healthComponent, uint32 eid, int startHealth) {
 	Component_Add(healthComponent, eid);
 	healthComponent->health[eid] = startHealth;
 }
