+++ src/HealthComponent.h
@@ -2,7 +2,7 @@
 #include "Component.h"
 
 struct HealthComponent : public Component {
-	uint8 health[Constants::MaxEntities_];
+	int health[Constants::MaxEntities_];
 };
 
-void HealthComponent_Add(HealthComponent* healthComponent, uint32 eid, uint8 startHealth);
+void HealthComponent_Add(HealthComponent* healthComponent, uint32 eid, int startHealth);
